move batch save to an extension method and simplify CanvasPaintingResolver
